+++ tests/iamnotroot.c
@@ -1,5 +1,5 @@
 /*
- * Copyright 2021-2023 Gaël PORTAY
+ * Copyright 2021-2024 Gaël PORTAY
  *
  * SPDX-License-Identifier: LGPL-2.1-or-later
  */
@@ -28,4 +28,8 @@ ssize_t path_resolution(int dfd, const char *path, char *buf, size_t bufsiz,
 	(void)atflags;
 
 	return __set_errno(ENOSYS, -1);
+}
+
+void unused()
+{
 }