+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [unreleased]
+
+- Intercept the functions `dladdr()` and `dladdr1()`
+
 ## [v23] - 2024-06-25
 
 ### Added